Split case relationships test into its own file
